requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
install==1.3.5
Markdown==3.4.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
gunicorn==20.1.0